Modify topbar to say either Log In or My Profile when appropriate
