+++ portfolio/src/main/webapp/profile.html
@@ -11,6 +11,7 @@
     <script src="topbar.js"></script>
     <div id="content">
       <h1 id="heading"></h1>
+      <div id="login-container"></div>
       <div id="character-distribution"></div>
       <div id="wordlength-distribution"></div>
       <div id="voting-piechart"></div>
