+++ portfolio/src/main/webapp/topbar.js
@@ -16,28 +16,30 @@ createTopBar();
 
 /** Creates a navigation bar and adds it to the page. */
 function createTopBar() {
-  	let topbar = document.getElementsByClassName("topbar")[0];
-    
-    let name = document.createElement("div");
-    name.setAttribute("class", "name");
-    name.innerText = "Lawrence Chen";
-    topbar.appendChild(name);
+    fetch('/login-data').then(response => response.json()).then(data => {
+        let topbar = document.getElementsByClassName("topbar")[0];
+        
+        let name = document.createElement("div");
+        name.setAttribute("class", "name");
+        name.innerText = "Lawrence Chen";
+        topbar.appendChild(name);
 
-    let menu = document.createElement("span");
-    menu.setAttribute("class", "menu");
-    menu.appendChild(createLink("index.html", "Home"));
-    menu.appendChild(createLink("aboutme.html", "About Me"));
-    menu.appendChild(createLink("projects.html", "Projects"));
-    menu.appendChild(createLink("https://www.linkedin.com/in/lawrence-chen-6ab742158/", "LinkedIn"));
-    menu.appendChild(createLink("https://github.com/ldcWV", "GitHub"));
-    menu.appendChild(createLink("resume.pdf", "Resume"));
-    topbar.appendChild(menu);
+        let menu = document.createElement("span");
+        menu.setAttribute("class", "menu");
+        menu.appendChild(createLink("index.html", "Home"));
+        menu.appendChild(createLink("aboutme.html", "About Me"));
+        menu.appendChild(createLink("projects.html", "Projects"));
+        menu.appendChild(createLink("https://www.linkedin.com/in/lawrence-chen-6ab742158/", "LinkedIn"));
+        menu.appendChild(createLink("https://github.com/ldcWV", "GitHub"));
+        menu.appendChild(createLink("resume.pdf", "Resume"));
+        topbar.appendChild(menu);
 
-    let otherlinks = document.createElement("span");
-    otherlinks.setAttribute("class", "otherlinks");
-    otherlinks.appendChild(createLink("login.html", "Log in"));
-    otherlinks.appendChild(createLink("profile.html", "My Profile"));
-    topbar.appendChild(otherlinks);
+        let otherlinks = document.createElement("span");
+        otherlinks.setAttribute("class", "otherlinks");
+        if(data.loggedIn) otherlinks.appendChild(createLink("profile.html", "My Profile"));
+        else otherlinks.appendChild(createLink("profile.html", "Log In"));
+        topbar.appendChild(otherlinks);
+    });
 }
 
 function createLink(url, text) {
